Update command framework to use game.Game as command handler
This leaves the implementation of /give incomplete, with a TODO item. Since we no longer have access to the internal player structures, we should do what the notchian server does and spawn the items in the player's chunk so they can pick it up. Alternatively, we can revisit the privacy of player.

Tests have been updated, and do pass.